break up course feature and consolidate global types
