+++ src/components/course/CourseBadge.tsx
