+++ src/lib/types/course.ts
