+++ src/lib/types/index.ts
@@ -1,0 +1,4 @@
+// barrel file for types
+
+export * from "@/lib/types/course";
+export * from "@/lib/types/planner";
